Move parsing into its own directory (#926)
* moving all parsing functions into parse.cpp

* making configuration its own directory

* fixing includes

* correcting python includes

* fix javascript
