+++ include/musica/configuration/chemistry.hpp
