+++ include/musica/miam/miam_builder.hpp
@@ -6,7 +6,7 @@
 #pragma once
 
 #include <musica/miam/miam_types.hpp>
-#include <musica/micm/chemistry.hpp>
+#include <musica/configuration/chemistry.hpp>
 #include <musica/micm/micm.hpp>
 #include <musica/utils/util.hpp>
