+++ include/musica/micm/micm_c_interface.hpp
@@ -5,9 +5,9 @@
 // It also includes functions for creating and deleting MICM instances with c bindings.
 #pragma once
 
-#include <musica/micm/chemistry.hpp>
+#include <musica/configuration/chemistry.hpp>
 #include <musica/micm/micm.hpp>
-#include <musica/micm/parse.hpp>
+#include <musica/configuration/parse.hpp>
 #include <musica/micm/solver_parameters.hpp>
 #include <musica/micm/state.hpp>
 #include <musica/utils/util.hpp>
