+++ javascript/bindings/musica_wasm.cpp
@@ -5,7 +5,7 @@
 
 #include <musica/micm/micm.hpp>
 #include <musica/micm/micm_c_interface.hpp>
-#include <musica/micm/parse.hpp>
+#include <musica/configuration/parse.hpp>
 #include <musica/micm/solver_parameters.hpp>
 #include <musica/utils/util.hpp>
 #include <musica/version.hpp>
