+++ src/configuration/CMakeLists.txt
@@ -1,0 +1,3 @@
+target_sources(musica PRIVATE
+  parse.cpp
+)
