+++ src/micm/micm_c_interface.cpp
@@ -1,6 +1,6 @@
 #include <musica/micm/cuda_availability.hpp>
 #include <musica/micm/micm_c_interface.hpp>
-#include <musica/micm/parse.hpp>
+#include <musica/configuration/parse.hpp>
 #include <musica/utils/error_handler.hpp>
 
 namespace musica
